refactor: extract extractSingleResume and pagination helpers into DI factories
Extract 4 functions from content.ts into dedicated factories:
- createResumeExtractor(deps) in lib/resume-extractor.ts
  - extractSingleResume (~300L)
- createPaginationUtils(deps) in lib/pagination-utils.ts
  - getPaginationInfo (~150L)
  - getNextPageButtonState (~50L)
  - waitForPagination (~50L)

Changes:
- Create lib/resume-extractor.ts with full extractSingleResume implementation
- Create lib/pagination-utils.ts with 3 pagination helpers
- Wire both factories in content.ts with proper dependency injection
- Remove inline function definitions from content.ts
- Update content.ts line count: 2,104L (-284L from main)

Follows established DI factory pattern with injected doc/win and
dependency objects for testability and modularity.
